functions/cloud-function-docai-from-bucket-sync/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
google-api-core
google-auth==2.2.0
google-cloud-core
google-cloud-bigquery
google-cloud-documentai
google-cloud-firestore
google-cloud-storage
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==3.2
packaging==21.0
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.7